Decode cloud file sizes as floats
alpha.11 broke every cache lookup on a real appliance:

  failed to inspect cached Morpheus virtual image "omni-talos-..."
  (id 215): json: cannot unmarshal number 0.21691131591796875 into
  Go struct field VirtualImageFile.cloudFiles.size of type int64

Morpheus reports a cloud file's size in fractional gigabytes, with
contentLength carrying the byte count. The Morpheus CLI's source says
exactly that and the struct was still typed as bytes. Both are now
floats: neither is relied on, and a number is a number.

The regression test uses the payload shape a live appliance returns.

+++ internal/pkg/provider/api.go
@@ -114,12 +114,17 @@ type VirtualImage struct {
 }
 
 // VirtualImageFile is one file held for a virtual image.
+//
+// Only the name is relied on. The sizes are decoded as floats because
+// Morpheus reports Size in fractional gigabytes (0.2169... for a Talos
+// qcow2), and a stricter type here failed every cache lookup on a real
+// appliance.
 type VirtualImageFile struct {
 	Name string `json:"name"`
-	// ContentLength is the size in bytes on current appliances; older ones
-	// report Size instead. Either may be zero when the other is set.
-	ContentLength int64 `json:"contentLength"`
-	Size          int64 `json:"size"`
+	// ContentLength is the size in bytes; Size is the same in gigabytes,
+	// and the only one of the two older appliances report.
+	ContentLength float64 `json:"contentLength"`
+	Size          float64 `json:"size"`
 }
 
 // HasFile reports whether Morpheus holds at least one file for the image.
